@@ -1,6 +1,9 @@
+'use client'
+
 import { useQuery, useQueryClient } from '@tanstack/react-query'
+import Image from 'next/image'
 import Link from 'next/link'
-import { useRouter } from 'next/router'
+import { usePathname } from 'next/navigation'
 
 import type { ITrackHeader } from '../../lib/interfaces/track-header'
 import { getDoesUserHaveAlbumSaved } from '../../lib/spotify'
@@ -11,7 +14,7 @@ import { SaveAlbumButton } from '../button'
 
 export default function TrackHeader({ data }: ITrackHeader) {
   const queryClient = useQueryClient()
-  const { asPath } = useRouter()
+  const pathname = usePathname()
 
   const fetchDoesUserHaveAlbumSaved = async () => {
     const isAlbumSaved = await getDoesUserHaveAlbumSaved(data.id)
@@ -26,10 +29,13 @@ export default function TrackHeader({ data }: ITrackHeader) {
 
   return (
     <div className="mt-5">
-      <img
+      <Image
         src={data.images.length && data.images[0] ? data.images[0].url : '/images/nocover.webp'}
-        className="mx-auto h-80 w-80 rounded-md object-cover"
         alt={data.name}
+        width="0"
+        height="0"
+        sizes="100vw"
+        className="mx-auto h-80 w-80 rounded-md object-cover"
       />
 
       <div className="mx mt-5 flex flex-col items-center gap-y-2">
@@ -43,9 +49,12 @@ export default function TrackHeader({ data }: ITrackHeader) {
 
         <div className="h-5">{isAlbumSaved && <SaveAlbumButton id={data.id} saved={isAlbumSaved[0]} />}</div>
 
-        {asPath == `/playlists/${data.id}` && data.tracks.total >= 5 && (
-          <Link href={`/recommendations/${data.id}`}>
-            <a className="my-5 max-w-fit cursor-pointer rounded-full bg-cyan-600 py-2 px-5 font-semibold text-white">Get recommendations</a>
+        {pathname == `/playlists/${data.id}` && data.tracks.total >= 5 && (
+          <Link
+            href={`/recommendations/${data.id}`}
+            className="my-5 max-w-fit cursor-pointer rounded-full bg-cyan-600 py-2 px-5 font-semibold text-white"
+          >
+            Get recommendations
           </Link>
         )}
 
