@@ -1,3 +1,7 @@
+'use client'
+
+import Image from 'next/image'
+
 type Props = {
   image: string
   alt: string
@@ -5,6 +9,14 @@ type Props = {
 }
 
 export default function CardImage({ image, alt, rounded }: Props) {
-  console.log(alt)
-  return <img src={image} className={`mb-5 object-cover ${rounded ? 'h-32 w-32 rounded-full' : 'h-48 w-48 rounded-md'}`} alt={alt} />
+  return (
+    <Image
+      src={image}
+      alt={alt}
+      width="0"
+      height="0"
+      sizes="100vw"
+      className={`mb-5 w-full object-cover ${rounded ? 'h-32 w-32 rounded-full' : 'h-48 w-48 rounded-md'}`}
+    />
+  )
 }
