@@ -1,3 +1,5 @@
+'use client'
+
 import Link from 'next/link'
 
 import type { IArtistsAlbums } from '../../lib/interfaces/artist-album'
@@ -15,20 +17,16 @@ export default function AlbumGrid({ items }: IArtistsAlbums) {
                 <li key={album.id}>
                   <CardItem>
                     <Link href={`/albums/${album.id}`}>
-                      <a>
-                        <CardImage
-                          image={album.images.length && album.images[1] ? album.images[1].url : '/images/nocover.webp'}
-                          alt={album.name}
-                        />
+                      <CardImage
+                        image={album.images.length && album.images[1] ? album.images[1].url : '/images/nocover.webp'}
+                        alt={album.name}
+                      />
 
-                        <CardName name={album.name} />
-                      </a>
+                      <CardName name={album.name} />
                     </Link>
 
-                    <Link href={`/artists/${album.artists[0].id}`}>
-                      <a className="hover:underline hover:decoration-white">
-                        <CardInfo info={album.artists[0].name} />
-                      </a>
+                    <Link href={`/artists/${album.artists[0].id}`} className="hover:underline hover:decoration-white">
+                      <CardInfo info={album.artists[0].name} />
                     </Link>
                   </CardItem>
                 </li>
