@@ -1,4 +1,7 @@
+'use client'
+
 import { useQuery } from '@tanstack/react-query'
+import Image from 'next/image'
 
 import type { IArtistHeader } from '../../lib/interfaces/artist-header'
 import { getDoesUserFollowArtist } from '../../lib/spotify'
@@ -19,10 +22,13 @@ export default function ArtistHeader({ data }: IArtistHeader) {
 
   return (
     <div className="text-center">
-      <img
+      <Image
         src={data.images.length && data.images[0] ? data.images[0].url : '/images/nocover.webp'}
-        className="mx-auto h-96 w-96 rounded-md object-cover"
         alt={data.name}
+        width="0"
+        height="0"
+        sizes="100vw"
+        className="mx-auto h-96 w-96 rounded-md object-cover"
       />
       <div className="mt-8 text-3xl font-black text-white md:text-6xl lg:text-8xl">{data.name}</div>
 
